% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lireZiff.R
\name{categorieEngin}
\alias{categorieEngin}
\title{Détermine une catégorie d'engin de pêche selon les numéros d'engins.}
\usage{
categorieEngin(x)
}
\arguments{
\item{x}{un vecteur de numéros d'engins à classifier en catégories}
}
\value{
une table en trois colonnes des catégories d'engin, soit le nom du type d'engin, l'étiquette en francais et l'étiquette en anglais
}
\description{
Détermine une catégorie d'engin de pêche selon les numéros d'engins.
}
