% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lireZiff.R
\name{lireZiff}
\alias{lireZiff}
\title{Lit les fichiers ZIFF}
\usage{
lireZiff(
  no_espece = NULL,
  no_espece_visee = NULL,
  annees = NULL,
  dir_input = file.path("//ent.dfo-mpo.ca", "dfo-mpo", "GROUP", "QUE", "Reg_Shares",
    "DFO", "science", "DAISS", "BD_Peches", "Ziff", "Version_totale"),
  dir_sauvegarde_locale = NULL,
  ne_pas_mettre_a_jour_version_locale = FALSE
)
}
\arguments{
\item{no_espece}{un codes STRAP de l'espèce capturée par la pêche. NULL importe toutes les espèces capturées.}

\item{no_espece_visee}{un codes STRAP de l'espèce visée par la pêche. NULL importe toutes les espèces visées.}

\item{annees}{un vecteur des années à considérer.}

\item{dir_input}{chemin du répertoire à utiliser pour lire les données telles que produites par la DAISS.}

\item{dir_sauvegarde_locale}{chemin du répertoire de sauvegarde local, si désiré. Permet de réduire le temps de lecture en ne lisant que les fichiers nécessitant une mise à jour.}

\item{ne_pas_mettre_a_jour_version_locale}{si TRUE, utiliser la version présente à \code{dir_sauvegarde_locale} sans mettre à jour les fichier. Utile pour reproduire des calculs antérieurs.}
}
\value{
un \code{data.frame} des ziff lus
}
\description{
\code{lireZiff()} lit les fichiers \verb{Zonal Interchange File Format} (ziff) préparés par la DAISS et disponbles sur le réseau interne, qui regroupe par année les données ziff des 4 régions
de l'est canadien, pour toutes les espèces.
Certaines variables sont ajusté pour des besoins de standardisation avec d'autres jeux de données.
}
\details{
Certains codes d'espèce fréquement utilisés: 100(morue), 120(sébaste sp.), 117(sébaste atlantique/du nord), 118(sébaste acadien/Rose),
130(flétan atlantique), 144(turbot), 140(plie can), 142(plie grise), 141(limande), 143(plie rouge), 702(crevette)
Pour une liste détaillée des codes d'espèce, voir /DCQCIMLNA01A/BD_Peches/Ziff/Documentation/Codes_espèces.xlsx
Les catégories de longueur utilisent des borne inférieurs inclusives et des bornes supérieurs exclusives, avec une tolérence de 0.01cm.
Ainsi, "plus grand que 85cm" devient >=84.99, la tolérence permettant d'éviter des comportements imprévisibles (et parfois incompris...) de la fonction.

Certains utilisateurs suggèrent d'installer les fichiers .txt localement pour accélérer le processus.
Cette version de la fonction offre la possibilité d'enregistrer localement le resultat de la lecture, et par la suite ne lire que les fichier mis à jour.

Note: À faire, valider si les opano sont correct pour toutes les années. Valider les "4Ru" si on met dans "4R"
}
